Refactoring popper to popup
